+++ lib/text_utils/text_utils.h
@@ -1,9 +1,7 @@
 #ifndef TEXT_UTILS_H
 #define TEXT_UTILS_H
 
-#include <stdint.h>
-
-void trimStart(char* text, uint8_t characters);
+void trimStart(char* text, size_t characters);
 size_t stringToSizeT(char* text);
 
 #endif
